pyproject.toml
[tool.poetry]
name = "teleser"
version = "0.1.2"
description = "A Telegram bot for remote command execution with optional user authorization, customizable command validation, and customizable messages."
authors = ["kai <thuan.dv0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.4"
nest-asyncio = "^1.6.0"


[tool.poetry.group.dev.dependencies]
coverage = "^7.6.1"
flake8 = "^7.1.1"
pre-commit = "^3.8.0"
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
